refactor(#430 W4): four copies of the timer entry, and the clock reached one
Phase-425 added the clock axis to `Executor::register_timer_on_clock` and to
nothing else, because there was nowhere else to add it to: `register_timer`,
`register_timer_on`, `register_timer_oneshot` and `register_timer_on_clock`
were four near-identical bodies writing a `TimerEntry` into the arena, and a
fifth axis lands in whichever one the author was looking at. That is how a
capability ends up reachable from the executor and from no node-level surface.

One body now — `register_timer_entry(node_id, period, source, oneshot, group,
callback)` — and the four public verbs are its argument lists. Two behavioural
non-changes, both checked rather than assumed:

  * `register_timer` and `register_timer_oneshot` now reach
    `apply_node_default_sched(slot, None, None)`, which returns on its first
    line when `node_id` is `None`.
  * every entry now seeds `last_clock_ns: source.now_ns()`, which is 0 for
    `Steady` — the literal the three wall verbs wrote.

`TimerClockSource` also MOVES, from `executor::arena` to the ungated
`crate::timer`. `mod arena` is `#[cfg(any(has_rmw, test))]`, so a build with no
transport cannot name the enum — and the declarative metadata layer
(`nros::node_metadata`, compiled in metadata mode with no transport) is exactly
such a build. W4 needs it to record which clock a declared timer asked for, and
an enum the metadata layer cannot name is an enum the metadata layer re-invents:
the 0135/0160 drift, one layer up. `executor` re-exports it and `arena` keeps a
`pub(crate) use`, so every path phase-425 wrote still resolves, including
`super::arena::TimerClockSource` in the sim-time test. The `clock()`/`now_ns()`
helpers stay `cfg(any(has_rmw, test))` — only the arena reads a clock, and
without a backend they would be dead code under `-D warnings`.

Sweep: `grep -rn 'fn create_timer\|fn register_timer' packages/api packages/core`
and `grep -rn 'TimerEntry {' packages` — one production construction site left.
